Create skeleton HTML document
This `index.html` is the base for the website we're going to build.

It starts with the basic HTML tag and the `<head>` tag, which contains
some important information to tell the browser how to display the page.
It also includes a space to put the `<title>` of our page, which will
appear at the top of the web browser window.

The `<body>` tag is where all the action happens. Inside this tag, we
have a `<div>` tag. `<div>` tags are block elements, as opposed to
inline elements. Inline elements flow across a page like text, and can
have line breaks etc., while block elements behave as if they're inside
a box, and everything needs to work around them.

This `<div>` tag will be used as the "container" for our web page. This
is so that later on we can write code to describe how our content should
behave, and apply it to this `<div>`. Note also that the `<div>` has an
ID of "div_content". In a web page, each ID must be unique, meaning that
if we ever need to refer to this `<div>`, we can use the ID
'div_content', and the browser will know what we mean.

Inside 'div_content', we have three `<section>` elements. These sections
are for our "header", "content", and "footer".

We've added in some elements that look `<!-- like this -->`. These are
HTML "comment" elements. The browser ignores them, but they are useful
to us to explain what we're doing. Some of the comments we've added look
`<!--x like this x-->`, with `x` characters at either end. These are
just regular comments, except that we've made them look a little
different so we can refer to them later, and we want to ensure we can
distinguish them from the rest of our comments.

Looking at the comments, you can see the bits that make clear where the
"header", "content", and "footer" of our document are.

There are also various other comments in there that are acting as
place-holders for more content we'll be adding later on.

In our content section, we have three separate code blocks:
- div_first_content: Here we'll be putting an image and some text.
- div_second_content: Here we'll be putting some lists and more text
- div_third_content: Here we'll be putting a table and a link to
  another page

This file looks big and complicated, but when you look at it in small
chunks, it should make sense.