Add heading and description of page
Let's start by adding a heading and a paragraph of content.

First, we'll set the `<title>` of our page to the name of our Patrol.
I'm going to set mine to "Scorpion Patrol".

Then, down in the "div_headings_paragraphs" div, I'm creating a heading
for my page, as well as a paragraph to explain what it's all about.