Create a navigation menu linking to a second page
Let's create a hyperlink.

Under the "Put your menu links here" comment, create an unordered list,
with two items:
1. "Home", linking to "/"
2. "My Page", linking to "my-page.html"

Note that we use "/" instead of "index.html". This is because by
default, when we browse to a folder (rather than a file) using a web
browser, the web server automatically looks for "index.html" as the file
to load, so we don't need to specify it manually.

This link should now go to the "my-page.html" file we renamed earlier
on. The other thing we'll do with this link is set a `target="_blank"`
attribute. This tells the browser to open this link in a new tab.

You might also want to edit the "my-page.html" file with some links to
get back to the home page. There's an example below.