Inline CSS Styles
While HTML is used to define our website *content*, CSS,
or Cascading Style Sheets, are used to style the layout and appearance
of our site.

CSS styles can live in three places. The first type we'll see is an
"inline style"

Inline styles are stored in the "style" attribute of a HTML element. If
we had a paragraph like this:

```html
<p>Hello world</p>
```

And we wanted to give it some colour, we could add a "style" attribute
like this:

```html
<p style="color: red">Hello world</p>
```

Let's change the colour of the "Our Patrol Members" heading, and the
font of the "Our Top Activities".