Switch to external stylesheet
Because of the way CSS works, it makes sense to apply it to *multiple*
elements at once, rather than one at a time.

For example, maybe we want *all* our headings to be orange. If we were
to do this, we wouldn't want to use an inline style for every `<h3>` in
our page.

Instead we use something known as "CSS selectors". CSS Selectors are a
way of saying "apply this style to all headings, or all paragraphs, or
this specific item or type of element."

To do this though, we need to move our CSS from "inline" to somewhere
else.

Aside from "inline", the two other CSS style formats are "internal", and
"external".

Internal CSS lives usually at the top of our HTML file You can see below
that we have a `<style>` element at the top of our HTML file just before
the end of the `<head>` element. Anything inside this `<style>` tag can
be applied to relevant elements in the same document.

```diff
     <title>Scorpion Patrol</title>
+    <style>
+        .heading_a {
+            color: orange;
+        }
+        .heading_b {
+            font-family: Impact;
+        }
+    </style>
 </head>
```

We're going to skip over "internal" CSS, and move straight to "external"
CSS. An external CSS style lives in a file all of its own.

Let's create a folder called "css" inside our "static" folder, and then
inside the "css" folder, create a file called "style.css"

Inside that "style.css" file, we put our CSS stylesheet content.

```css
.heading_a {
    color: orange;
}
.heading_b {
    font-family: Impact;
}
```

This file is then referenced inside any HTML file that wants to use it,
by adding a `<link>` tag to the `<head>` of the document:

```html
<link href="static/css/style.css" rel="stylesheet">
```

The advantage of this is that we can keep our style completely separate
from our content, *and* the same style can be used on multiple pages.

So let's break down how this style works now that it's not inline.

We mentioned CSS selectors before. CSS selectors can be used to select a
group of elements by element *type* (e.g. "all `<p>` elements), or we
can reference our elements ourselves by assigning them a "class" or "id"
attribute.

For example, we've assigned the `class="heading_a"` attribute to one of
our headings. Multiple elements in our document can have a "heading_a"
class, and any CSS rule that references "heading_a" will apply to those.

To reference a class in CSS, we start it with a dot (`.`):
```css
.heading_a { color: orange; }
```

If we just wanted to refer to all `<h3>` elements, we'd just say `h3`
with nothing else:
```css
h3 { color: orange; }
```

Finally, we could have a heading with a specific ID attribute:
```html
<h3 id="my_special_heading">Special heading!</h3>
```

Unlike classes, which can be reused, only one element in a whole HTML
file can have this ID.

We can reference this element in CSS by starting it with a hash (`#`):
```css
 #my_special_heading { color: orange; }
```