More colours and fonts
In HTML and CSS, there a several different ways we can refer to a colour
that we want to use. The most straightforward is to use the colour's
keyword.

A colour keyword is something like "blue" or "green". There are quite a
few of both [basic](https://www.w3.org/TR/css-color-3/#html4) and
[extended](https://www.w3.org/TR/css-color-3/#svg-color) colour
keywords, but there are a *lot* of colours out there, so we can't name
them all.

The most common way to refer to colours is by how they are made up. When
it comes to colours, computers think in terms of red, green, and blue.
By using different brightnesses of these three colours, we can make up
most other colours. Each of these three colours can have a value between
0 and 255. So, for example, Cub Scout Yellow is 255 red, 201 green, and
12 blue. We can write this as `RGB(255, 201, 12)` (where RGB stands for
Red, Green, Blue):

```html
<p style="color: RGB(255, 201, 12)">Hello world</p>
```

<p style="color: RGB(255, 201, 12)">Hello world</p>

Often though, computers count by 16s instead of 10s, using a number
system called "hexadecimal" (our usual number system is called
"decimal"). We won't cover it in great detail, but all that you really
need to know right now is that in hexadecimal, there are 16 "numbers"
instead of 10:

|Hexadecimal|Decimal|
|:-:|-:|
|0|0|
|1|1|
|2|2|
|3|3|
|4|4|
|5|5|
|6|6|
|7|7|
|8|8|
|9|9|
|A|10|
|B|11|
|C|12|
|D|13|
|E|14|
|F|15|

Also, when we write colours in hexadecimal, we start them with a hash
(`#`; type this with <kbd>Shift</kbd>+<kbd>3</kbd>). So Cub Scout Yellow
instead looks like `#ffc90c`:

```html
<p style="color: #ffc90c">Hello world in hexadecimal</p>
```

<p style="color: #ffc90c">Hello world in hexadecimal</p>

Let's use some different types of colours for our headings.

While we're at it, we'll add "heading_c" class to the
"Our Favourite Campsite" heading, and add some colour and a different to
font to that as well.