Add a style to headings in only one section
Let's look at using an element's ID to style it.

At the same time, we're going to do something else new: we're going to
use *two* separate CSS selectors in one rule.

This rule uses an ID (you can tell because it starts with a `#`),
followed by a space, and then a `h3`. This says: "for *any* h3 that is
*inside* an element with this ID, apply this style.