Adding Bootstrap
Writing CSS styles manually is all well-and-good, but there are tools
that can make things even easier for us.

Bootstrap is a library of CSS styles that we can add to our web page
that will not only automatically style it to look a bit nicer without us
doing anything, but it gives a huge collection of CSS classes we can add
to our elements to change how they look.

Let's start by just adding the Bootstrap stylesheet to our document and
taking a look at what it does.

Take care to add the Bootstrap immediately *before* our own CSS
stylesheet. If we add it *after* our own file, we won't be able to use
our styles to override Bootstrap if we wish.

Note that this link we're adding is to 'cdn.scouthack', which is a
service running on our ScoutHack server. After the weekend, this could
be pointed to a copy of Bootstrap that is out on the internet so that it
can be used from anywhere.