Using containers
Let's add `class="container"` to our `div_content`. A Bootstrap
container constrains the width of our page so that it is easier to read.

After doing this, we'll see that no matter how wide our browser window
is, the content has a maximum width that is easier on the eyes.