Add a Bootstrap row and columns
Bootstrap can think of a webpage as having rows that are each divided
into 12 equal columns.

We add a `row` class to our `div_first_content` div, and then add
a `col-3` class to our `div_patrol_logo`and a `col-9` class to our
`div_headings_paragraphs` div.

This tells Bootstrap to make the patrol logo take up 3/12 of our screen,
and the heading and paragraph to take up the remaining 9/12 of the
screen.