More rows and columns
Let's do the same with our other divs.

For the `div_second_content`, we split the three components into equal
thirds, and for the `div_third_content` we use 9/12 for our table, and
3/12 for our other section.

Also for the second div we're adding a 'py-5' class, which says to add 5
units of padding (the "p") to the top and bottom (the "y" axis) of this
div.