Change body background and button colour
Even though we're using Bootstrap, that doesn't mean we can't still
manually edit our own custom CSS stylesheet.

Let's set our body's background colour. If we were to *only* do this,
that colour setting would be inherited by everything *inside* the body,
which is not what we want.

To counteract this, we re-set our `container` class to have a white
background.

Also, because our own stylesheet appears *after* the Bootstrap
stylesheet in our web page, we can override the colour of `btn-primary`
here.