Update links in index
Flask has a special way to reference files in your HTML. To ensure Flask
can find resources properly, we'll update our links to Flask's way of
doing things.

Flask uses a template language called Jinja. Jinja lets you
dynamically add content to your HTML files using programming. A Jinja
"expression" appears between two pairs of `{{ curly brackets }}`.

Inside the curly brackets, we can call Python functions whose output
will appear in our web page. We'll be using the `url_for` function in
two ways. For the first, we pass it the `'static'` argument, and a
`filename` argument to tell it we're looking for a file in our "static"
folder.

For the second, we'll be directly referring to the function we created
in our `website.py`.

By complete fluke, because we were already using a static folder, our
links continued to work after our transition to Flask, but it's still
best practice to make this change.