Replace block comments with proper Jinja template blocks
Earlier on we pointed on that we'd added some special HTML comments that
looked `<!--x like this x-->` with `x`es in them. Now we're going to
make use of those.

We're going to do a find-and-replace to replace them with the curly
braces and percent signs that Jinja uses for its "statements" in our
templates.

When we're done, anything that looked `<!--x like this x-->` originally
will now look `{% like this %}`.

Here's a helpful `sed` command to do that replacement:

```
sed -i -e 's/<!--x/{%/g' -e 's/x-->/%}/g' templates/index.html
```