Break out index into base and patrol templates
Now that we've put some blocks in our HTML, let's split it out to
properly use it as a template.

We'll copy `templates/index.html` to both `templates/base.html` *and*
`templates/patrol.html`.

For `templates/base.html`, we'll remove everything inside the "content"
block, and the "title" block.

For the `templates/patrol.html`, we'll remove everything *outside* of
the "title" and "content" blocks.

Finally, we'll update the template referenced in `website.py` from
`index.html` to `patrol.html`.

After a refresh, it should *look* like nothing has changed.