Add page to display submitted program ideas
The final thing we want is to be able to see the program ideas that
users have submitted.

We'll create a new route which reads from our CSV file and compiles a
list of authors and ideas. This route will then send that list to a
newly created `program.html` template to be displayed.

Now we've got this page, we'll update our links so that this page shows
up after people submit a new program idea, and we'll add a link to the
page from our Patrol page.